backend/adk_host/requirements.txt
# ADK Host Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx>=0.28.1
python-dotenv==1.0.0